+++ src/plugins/lcm_plugin/lcm_plugin.cc
@@ -59,7 +59,8 @@ void LcmPlugin::Shutdown() noexcept {
 }
 
 void LcmPlugin::SetPluginLogger() {
-  SetLogger(aimrt::logger::LoggerRef(core_ptr_->GetLoggerManager().GetLoggerProxy(Name()).NativeHandle()));
+  std::string log_module_name = "core/" + std::string(Name());
+  SetLogger(aimrt::logger::LoggerRef(core_ptr_->GetLoggerManager().GetLoggerProxy(log_module_name).NativeHandle()));
 }
 
 void LcmPlugin::RegisterLcmRpcBackend() {