+++ src/plugins/mqtt_plugin/mqtt_plugin.cc
@@ -119,7 +119,8 @@ void MqttPlugin::Shutdown() noexcept {
 }
 
 void MqttPlugin::SetPluginLogger() {
-  SetLogger(aimrt::logger::LoggerRef(core_ptr_->GetLoggerManager().GetLoggerProxy(Name()).NativeHandle()));
+  std::string log_module_name = "core/" + std::string(Name());
+  SetLogger(aimrt::logger::LoggerRef(core_ptr_->GetLoggerManager().GetLoggerProxy(log_module_name).NativeHandle()));
 }
 
 void MqttPlugin::RegisterMqttChannelBackend() {