+++ src/plugins/net_plugin/net_plugin.cc
@@ -237,7 +237,8 @@ void NetPlugin::Shutdown() noexcept {
 }
 
 void NetPlugin::SetPluginLogger() {
-  SetLogger(aimrt::logger::LoggerRef(core_ptr_->GetLoggerManager().GetLoggerProxy(Name()).NativeHandle()));
+  std::string log_module_name = "core/" + std::string(Name());
+  SetLogger(aimrt::logger::LoggerRef(core_ptr_->GetLoggerManager().GetLoggerProxy(log_module_name).NativeHandle()));
 }
 
 void NetPlugin::RegisterHttpRpcBackend() {