+++ src/runtime/core/logger/rotate_file_logger_backend.h
@@ -28,7 +28,7 @@ class RotateFileLoggerBackend : public LoggerBackendBase {
   std::string_view Type() const override { return "rotate_file"; }
 
   void Initialize(YAML::Node options_node) override;
-  void Start() override;
+  void Start() override {}
   void Shutdown() override { run_flag_.store(false); }
 
   void RegisterGetExecutorFunc(