+++ document/release_notes.md
@@ -12,8 +12,8 @@
 - 优化了main_executor的性能；
 - 提供了rpc/channel统一的backend开关；
 - 提供了每个Module独立的enable开关；
-- 提供了同步RPC接口和异步RPC接口，以及相关example；
-- 原来的协程RPC Proxy接口重命名为CoProxy，之前老的命名将在后几个版本中删除；
+- 提供了同步和异步的RPC接口，以及相关example；
+- 原来的协程RPC Proxy/Service接口重命名为CoProxy/CoService，之前老的命名将在后几个版本中删除；
 - 优化了框架日志，提供【Initialization Report】，现在可以在启动日志中查看executor信息、channel/rpc注册信息；
 - consoler/file日志后端现在支持模块过滤功能，并且允许同时注册多个file日志后端，方便将不同模块的日志打印到不同文件中；
 - ros2_plugin：